requirements.txt
# Core dependencies
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1

# Deep learning
tensorflow>=2.6.0
keras>=2.6.0

# Image processing
opencv-python>=4.5.3
Pillow>=8.3.1

# Web application
flask>=2.0.1
flask-wtf>=0.15.1
flask-sqlalchemy>=2.5.1

# Database
sqlite3>=2.6.0

# Testing
pytest>=6.2.5